vector type changed from int to double, formatted (llvm style)

+++ homework/inner-product/main.cpp
@@ -5,10 +5,10 @@
 using namespace std;
 
 int main(void) {
-  vector<double> v1{1, 2, 3, 4};
-  vector<double> v2{1, 2, 3, 4};
-  vector<double> v3{7, 4, 3};
-  vector<double> v4{17, 6, 2};
+  vector<int> v1{1, 2, 3, 4};
+  vector<int> v2{1, 2, 3, 4};
+  vector<int> v3{7, 4, 3};
+  vector<int> v4{17, 6, 2};
   cout << "ArithmeticAverage(v1,v2): " << ArithmeticAverage(v1, v2) << endl;
   cout << "Distance(v3,v4): " << Distance(v3, v4) << endl;
   return 0;
